+++ examples/messaging/src/App.js
@@ -10,6 +10,7 @@ import {
   MessageSimple,
   ChannelHeader,
   ChannelPreviewMessenger,
+  InfiniteScrollPaginator,
   ChannelListMessenger,
   ChannelList,
   Window,
@@ -23,7 +24,7 @@ const urlParams = new URLSearchParams(window.location.search);
 const user =
   urlParams.get('user') || process.env.REACT_APP_CHAT_API_DEFAULT_USER;
 const theme = urlParams.get('theme') || 'light';
-const channelName = urlParams.get('channel') || 'demo';
+// const channelName = urlParams.get('channel') || 'demo';
 const userToken =
   urlParams.get('user_token') ||
   process.env.REACT_APP_CHAT_API_DEFAULT_USER_TOKEN;
@@ -41,60 +42,36 @@ class App extends Component {
       },
       userToken,
     );
-
-    this.channel = this.chatClient.channel('messaging', channelName, {
-      image:
-        'https://images.unsplash.com/photo-1512138664757-360e0aad5132?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2851&q=80',
-      name: 'The water cooler',
-      example: 1,
-    });
-    const exampleVersion = 1;
-    this.chatClient
-      .channel('messaging', `aww`, {
-        image:
-          'https://images.unsplash.com/photo-1425082661705-1834bfd09dca?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2855&q=80',
-        name: `aww`,
-        example: exampleVersion,
-      })
-      .watch();
-    this.channel.update({
-      image:
-        'https://images.unsplash.com/photo-1512138664757-360e0aad5132?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2851&q=80',
-      name: 'The water cooler',
-      example: exampleVersion,
-    });
-
-    this.channel.watch();
-
-    const filters = { type: 'messaging', example: 1 };
-    const sort = { last_message_at: -1 };
-    this.channels = this.chatClient.queryChannels(filters, sort, {
-      subscribe: true,
-    });
   }
 
   render() {
+    const filters = { type: 'messaging', example: 1 };
+    const sort = {
+      last_message_at: -1,
+      cid: 1,
+    };
+    const options = {
+      member: true,
+      watch: true,
+      limit: 30,
+    };
     return (
       <Chat client={this.chatClient} theme={`messaging ${theme}`}>
         <ChannelList
-          channels={this.channels}
           List={ChannelListMessenger}
           Preview={ChannelPreviewMessenger}
+          filters={filters}
+          sort={sort}
+          options={options}
+          Paginator={(props) => (
+            <InfiniteScrollPaginator threshold={300} {...props} />
+          )}
         />
-        <Channel
-          onMentionsHover={(e, user) => console.log(e, user)}
-          onMentionsClick={(e, user) => console.log(e, user)}
-        >
+        <Channel>
           <Window>
             <ChannelHeader />
             <MessageList TypingIndicator={TypingIndicator} />
-            <MessageInput
-              multipleUploads={false}
-              acceptedFiles={['image/*']}
-              maxNumberOfFiles={1}
-              Input={MessageInputFlat}
-              focus
-            />
+            <MessageInput Input={MessageInputFlat} focus />
           </Window>
           <Thread Message={MessageSimple} />
         </Channel>