+++ src/components/MessageLivestream.js
@@ -203,23 +203,26 @@ export class MessageLivestream extends React.PureComponent {
                 <span className={`str-chat__message-livestream-time`}>
                   {moment(message.created_at).format('h:mmA')}
                 </span>
-                {this.props.channelConfig.reactions && (
-                  <span onClick={this.onClickReactionsAction}>
+                {this.props.channelConfig &&
+                  this.props.channelConfig.reactions && (
+                    <span onClick={this.onClickReactionsAction}>
+                      <span
+                        dangerouslySetInnerHTML={{
+                          __html: reactionSvg,
+                        }}
+                      />
+                    </span>
+                  )}
+                {!this.props.threadList &&
+                  this.props.channelConfig &&
+                  this.props.channelConfig.replies && (
                     <span
                       dangerouslySetInnerHTML={{
-                        __html: reactionSvg,
+                        __html: threadSvg,
                       }}
+                      onClick={(e) => this.props.openThread(e, message)}
                     />
-                  </span>
-                )}
-                {!this.props.threadList && this.props.channelConfig.replies && (
-                  <span
-                    dangerouslySetInnerHTML={{
-                      __html: threadSvg,
-                    }}
-                    onClick={(e) => this.props.openThread(e, message)}
-                  />
-                )}
+                  )}
                 <span onClick={this.onClickOptionsAction}>
                   <span
                     dangerouslySetInnerHTML={{