+++ src/utils.js
@@ -163,7 +163,7 @@ export const renderText = (message) => {
     text = text.replace(urlInfo.raw, mkdown);
   }
   let newText = text;
-  if (mentioned_users.length) {
+  if (mentioned_users && mentioned_users.length) {
     for (let i = 0; i < mentioned_users.length; i++) {
       const username = mentioned_users[i].name || mentioned_users[i].id;
       const mkdown = `**@${username}**`;
@@ -193,3 +193,27 @@ export function generateRandomId() {
 function S4() {
   return (((1 + Math.random()) * 0x10000) | 0).toString(16).substring(1);
 }
+
+export const smartRender = (ElementOrComponentOrLiteral, props, fallback) => {
+  if (ElementOrComponentOrLiteral === undefined) {
+    ElementOrComponentOrLiteral = fallback;
+  }
+  if (React.isValidElement(ElementOrComponentOrLiteral)) {
+    // Flow cast through any, to make flow believe it's a React.Element
+    const element = ElementOrComponentOrLiteral;
+    return element;
+  }
+
+  // Flow cast through any to remove React.Element after previous check
+  const ComponentOrLiteral = ElementOrComponentOrLiteral;
+
+  if (
+    typeof ComponentOrLiteral === 'string' ||
+    typeof ComponentOrLiteral === 'number' ||
+    typeof ComponentOrLiteral === 'boolean' ||
+    ComponentOrLiteral == null
+  ) {
+    return ComponentOrLiteral;
+  }
+  return <ComponentOrLiteral {...props} />;
+};